refactor: reorganize code structure with proper separation of concerns
- Move validation schemas from model/ to schemas/ directory
- Separate business logic into models/ directory
- Update all imports to use new organized structure
- Improve code discoverability and maintainability
- Maintain all existing functionality with better organization

Changes:
- src/schemas/: JSON and Record validation schemas
- src/models/: Domain models and business logic
- Updated imports across all files
- Removed old model/ directory
