+++ src/models/index.ts
@@ -1,0 +1,2 @@
+// Domain models
+export * from './record'
