+++ src/schemas/index.ts
@@ -1,0 +1,5 @@
+// JSON schemas
+export * from './json'
+
+// Record schemas
+export * from './record'
