+++ src/service/record.service.ts
@@ -1,6 +1,7 @@
 import { JSONValue } from 'hono/utils/types'
-import { Record, RecordMetadata, RecordMetadataObject } from '../model/record'
-import { JSONPatch } from '../model/json'
+import { Record, RecordMetadata } from '../models/record'
+import { RecordMetadataObject } from '../schemas/record'
+import { JSONPatch } from '../schemas/json'
 import z4 from 'zod/v4'
 import { validateSchema } from '../lib/json-schema'
 import { HTTPException } from 'hono/http-exception'
